Re #15958 Start moving validate inputs into NexusProcessed
